requirements.txt
# Discord
discord.py>=2.3.0

# Database
supabase>=2.0.0
psycopg2-binary>=2.9.0

# Authentication & Security
PyJWT>=2.8.0
cryptography>=41.0.0
bcrypt>=4.0.0

# API Integration
aiohttp>=3.9.0
openai>=1.0.0
PyGithub>=2.1.0

# Monitoring & Logging
psutil>=5.9.0
prometheus_client>=0.17.0
structlog>=23.1.0

# Caching & Rate Limiting
redis>=5.0.0
aioredis>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0